refactor: decompose divisions and funding-programs detail pages
Extract type definitions, data parsing, and section components from
monolithic page files into focused modules.
